+++ Assets/Daemons Love & Carnage/Gameplay/Feature/HealthBar/Scripts/HealthBar.cs
@@ -5,12 +5,12 @@ public class HealthBar : MonoBehaviour
 {
     public Slider sliderBar;
 
-    public void MaxHealth(int health) //setta il valore massimo 
+    public void MaxHealth(float health) //setta il valore massimo 
     {
         sliderBar.maxValue = health;
     }
 
-    public void SetHealth(int health) //setta il valore della barra con la variabile intera
+    public void SetHealth(float health) //setta il valore della barra con la variabile intera
     {
         sliderBar.value = health;
     }
